+++ app/src/main/java/com/robandboo/fq/service/AnswerService.java
@@ -15,7 +15,7 @@ public interface AnswerService {
     Call<Answer> saveAnswer(@Body Answer answer);
 
     @GET("/api/v1/answer/getAnswerByQuestionId/{id}")
-    Call<List<Answer>> getAnswerByQuestionId(@Path("id") int id);
+    Call<List<Answer>> getAnswerByQuestionId(@Path("id") Long id);
 
     @POST("/api/v1/answer/save/file/vote/{id}")
     Call<Void> saveVote(@Path("id") Long id);
